Rename theme class folder
